// src/Constants/CreditManagementInstallmentInterval.d.ts
declare enum CreditManagementInstallmentInterval {
    DAY = "Day",
    TWODAYS = "TwoDays",
    WEEK = "Week",
    TWOWEEKS = "TwoWeeks",
    HALFMONTH = "HalfMonth",
    MONTH = "Month",
    TWOMONTHS = "TwoMonths",
    QUARTERYEAR = "QuarterYear",
    HALFYEAR = "HalfYear",
    YEAR = "Year"
}
export default CreditManagementInstallmentInterval;